♻️ 分離
