+++ src/app/_components/LoginButton/index.tsx
@@ -2,13 +2,10 @@ import IconBrandGitHub from "@/assets/brands/brand-github.svg";
 import IconBrandGoogle from "@/assets/brands/brand-google.svg";
 import Button from "@/components/Button";
 import ExternalLink from "@/components/ExternalLink";
-import { REDIRECT_TO_AUTH_ERROR } from "@/constants/redirect";
 import { links } from "@/constants/site";
-import { signIn } from "@/lib/auth";
-import { AuthError } from "next-auth";
-import { redirect } from "next/navigation";
 import type { ComponentPropsWithoutRef } from "react";
 import { twMerge } from "tailwind-merge";
+import { signInWithRedirect } from "./signIn";
 
 type Props = {
   /** ログイン後のリダイレクト先 */
@@ -17,26 +14,7 @@ type Props = {
 };
 
 export default function LoginButtons({ className, redirectTo }: Props) {
-  const handleSubmit = async (redirectUrl: string, formData: FormData) => {
-    "use server";
-    const provider = formData.get("provider")?.toString();
-
-    try {
-      await signIn(provider, {
-        redirectTo: redirectUrl,
-      });
-    } catch (error) {
-      // 認証のエラーならエラーページにリダイレクト
-      if (error instanceof AuthError) {
-        return redirect(REDIRECT_TO_AUTH_ERROR);
-      }
-
-      // それ以外のエラーはそのままthrow (Next.jsのリダイレクト等もここに流れる)
-      throw error;
-    }
-  };
-
-  const handleSubmitWithRedirect = handleSubmit.bind(null, redirectTo);
+  const handleSubmitWithRedirect = signInWithRedirect.bind(null, redirectTo);
 
   return (
     <div className={twMerge("w-full", className)}>
